+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## v0.1.2.0
 
+* Add `separateRecInput`, `separateRecInput'`, `stripRecInput`, `endRecInput`.
 * Inline everything.
 
 ## v0.1.1.0